chore: aws util v2
