+++ packages/collector/test/integration/currencies/cloud/aws-sdk/combined_products.js
@@ -14,23 +14,7 @@ const AWS = require('aws-sdk');
 const express = require('express');
 const logPrefix = `Combined AWS SDK v2 products (${process.pid}):\t`;
 
-function getLocalstackEndpoint() {
-  if (process.env.RUN_AWS === 'true') return null;
-  let endpoint = process.env.INSTANA_CONNECT_LOCALSTACK_AWS;
-  if (!endpoint) return null;
-  if (endpoint.startsWith('localstack://')) {
-    endpoint = endpoint.replace('localstack://', 'http://');
-  }
-  return endpoint;
-}
-
-function getClientConfig(extraOptions) {
-  const endpoint = getLocalstackEndpoint();
-  if (endpoint) {
-    return Object.assign({ region: 'us-east-2', endpoint, accessKeyId: 'test', secretAccessKey: 'test' }, extraOptions);
-  }
-  return { region: 'us-east-2' };
-}
+const { getClientConfig } = require('@_local/collector/test/integration/currencies/cloud/aws-sdk/aws-utils');
 
 AWS.config.update({ region: 'us-east-2' });
 const s3 = new AWS.S3(getClientConfig({ s3ForcePathStyle: true }));
